changed amp class handling
