+++ modules/Generic/amp_generic.py
@@ -2,6 +2,7 @@
 import AMP as AMP
 
 
+DisplayImageSources = ['Generic']
 class AMPGeneric(AMP.AMPInstance):
     def __init__(self, instanceID = 0, serverdata = {},Index = 0,Handler=None):
         super().__init__(instanceID,serverdata,Index,Handler=Handler)
